+++ ProlexNetUpdater/Library/Update/Update.cs
@@ -1,15 +1,11 @@
-﻿namespace ProlexNetUpdater.Library.Update
+﻿using ProlexNetUpdater.Library.Common;
+
+namespace ProlexNetUpdater.Library.Update
 {
     public class Update
     {
         public static void Run()
         {
-            //Vasculha registro pela pasta do ProlexNet             
-            Registry.LoadPath();
-
-            //Download da lista 
-            DownloadParameters.LoadApplicationList();
-
             //Pára o IIS
             IIS.Stop();
