+++ ProlexNetUpdater/ProlexNetUpdater.csproj
@@ -64,14 +64,14 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Constants.cs" />
-    <Compile Include="Library\Update\Download.cs" />
-    <Compile Include="Library\Update\DownloadParameters.cs" />
-    <Compile Include="Library\Update\Hash.cs" />
+    <Compile Include="Library\Common\Download.cs" />
+    <Compile Include="Library\Common\DownloadParameters.cs" />
+    <Compile Include="Library\Common\Hash.cs" />
     <Compile Include="Library\Update\IIS.cs" />
-    <Compile Include="Library\Update\Registry.cs" />
+    <Compile Include="Library\Common\Registry.cs" />
     <Compile Include="Library\Script\Script.cs" />
     <Compile Include="Library\Update\Update.cs" />
-    <Compile Include="Library\Update\Zip.cs" />
+    <Compile Include="Library\Common\Zip.cs" />
     <Compile Include="Models\ApplicationList.cs" />
     <Compile Include="Models\ProgramStatus.cs" />
     <Compile Include="Program.cs" />
@@ -91,6 +91,7 @@
       <Install>false</Install>
     </BootstrapperPackage>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\Costura.Fody.3.1.0\build\Costura.Fody.targets" Condition="Exists('..\packages\Costura.Fody.3.1.0\build\Costura.Fody.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
